% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_housing}
\alias{boston_housing}
\title{Boston housing data}
\format{A list of 4 components:
\describe{
  \item{train}{A \code{data.frame} of the training dataset which contains 354
  rows and 14 columns.}
  \item{test}{A \code{data.frame} of the test dataset which contains 152 rows
  and 14 columns.}
  \item{target_idx}{A column index of the target (response) varible.}
  \item{pred_idx}{A set of column indicies of the predictors.}
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Housing}
}
\usage{
boston_housing
}
\description{
This dataset concerns the values of 506 houses in suburbs of Boston.
}
\keyword{datasets}

