% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{german_credit}
\alias{german_credit}
\title{German credit data}
\format{A list of 4 components:
\describe{
  \item{train}{A \code{data.frame} of the training dataset which contains 700
  rows and 21 columns.}
  \item{test}{A \code{data.frame} of the test dataset which contains 300 rows
  and 21 columns.}
  \item{target_idx}{A column index of the target (response) varible.}
  \item{pred_idx}{A set of column indicies of the predictors.}
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Statlog+(German+Credit+Data)}
}
\usage{
german_credit
}
\description{
This dataset classifies 1,000 people described by a set of attributes as
good or bad credit risks.
}
\keyword{datasets}

